Recognize print request start and count anomalies
New code involves request.py which represents a newly detected request
and the contents of the outgoing message. runner.py does pick up
messages from two topics: the RAW_DATA_TOPIC which is crawled for a
rising set temperature signal, that indicates the start of a new print,
and events on the ANOMALY_TOPIC are counted towards the newly recognized
print.