Use generic module template
In the Hackathon midst of august we created some templates for
generic data processors. It already handles communication
(currently with kafka) and we just have to handle incoming events
via the EventHandler. Incoming messages come with the content august
dictionary and origin of the message as string (topic in this case).

The way it is coded now, the infrastructure and the use case
specific code are clearly seperated and can be tested more easily